Replace show --installed-deps with list
