+++ rye/src/cli/mod.rs
@@ -11,6 +11,7 @@ mod fmt;
 mod init;
 mod install;
 mod lint;
+mod list;
 mod lock;
 mod make_req;
 mod pin;
@@ -70,18 +71,9 @@ enum Command {
     Rye(rye::Args),
     Uninstall(uninstall::Args),
     Version(version::Args),
-    #[command(hide = true)]
     List(list::Args),
     #[command(hide = true)]
     Shell(shell::Args),
-}
-
-pub mod list {
-    /// There is no real list command yet.
-    ///
-    /// Use rye show --installed-deps instead
-    #[derive(clap::Parser, Debug)]
-    pub struct Args {}
 }
 
 pub mod shell {
@@ -136,11 +128,7 @@ pub fn execute() -> Result<(), Error> {
         Command::Rye(cmd) => rye::execute(cmd),
         Command::Uninstall(cmd) => uninstall::execute(cmd),
         Command::Version(cmd) => version::execute(cmd),
-        Command::List(..) => {
-            // until we have a proper list command, make it error with what the
-            // user should be using instead.
-            bail!("unknown command. Maybe you mean rye show --installed-deps");
-        }
+        Command::List(cmd) => list::execute(cmd),
         Command::Shell(..) => {
             bail!(
                 "unknown command. The shell command was removed. Activate the virtualenv instead with '{}' instead.",
